Network Connectivity -- HTTPS Request
A network request is made to fetch all list items. For verification, the
local database is populated with the request result.

Dependencies for the serialization library KotlinX Serialization. This
library simplifies parsing, serializing, and manipulation of JSON data.
Additionally, the library supports other serialization formats should
the format used for data changes.